fix database listing issue

+++ resources/views/home.blade.php
@@ -16,11 +16,13 @@
 
                     <ul class="list-group">
                         @foreach($database_list as $database)
-                          <li class="list-group-item d-flex justify-content-between align-items-center">
-                            <a href="{{url($database->name.'/details')}}"> {{$database->name}} </a>
-                            <span class="badge badge-primary">{{$database->db_size}}</span>
-                             <a onclick="return confirm('Are you sure you want to drop this database')" href="{{url('/drop_database/'.$database->name)}}"><span class="badge badge-danger">X</span></a>
-                          </li>
+                            @if(in_array($database->name,$list))
+                                <li class="list-group-item d-flex justify-content-between align-items-center">
+                                    <a href="{{url($database->name.'/details')}}"> {{$database->name}} </a>
+                                    <span class="badge badge-primary">{{$database->db_size}}</span>
+                                    <a onclick="return confirm('Are you sure you want to drop this database')" href="{{url('/drop_database/'.$database->name)}}"><span class="badge badge-danger">X</span></a>
+                                </li>
+                            @endif
                         @endforeach
                     </ul>
                 </div>
